requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
psycopg2-binary==2.9.9
redis==5.0.8
